feat: move process and clear logic to generic functions
